Virtual Shop - JEE+AngJS (2015)
The application implements a virtual store where users can buy a number
of products using web services.

Based in a REST architecture, the app combines a fresh interface using
technologies like HTML5, AngularJS and Bootstrap and the strength of
Java working over Spring, Hibernate and Maven for creating the back-end.